requirements.txt
# Core PyTorch (latest stable)
torch>=2.6.0
torchvision
torchaudio

# NLP & ML
transformers>=4.36.0
sentence-transformers>=2.2.0

# Vector Store
chromadb>=0.4.0
faiss-cpu>=1.7.4

# Code Processing
GitPython>=3.1.0
pygments>=2.17.0

# API
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
pyyaml>=6.0.0
pandas>=2.1.0

# Security - for vulnerability scanner
safetensors>=0.4.1
accelerate>=0.25.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0